Remove configuration frameworks and other non
fisherman is the best way to install and manage plugins in fish-shell, whether these plugins are in the fisherman ecosystem or in the oh-my-fish organization. 

Other frameworks like oh-my-fish, fundle and wahoo are no longer actively maintained and promote deprecated patterns like init.fish files, "hooks" and other non-standard mechanisms **no longer relevant** in fish-shell 2.3.0. 

Finally, awesome-fish should not aim to be another link farm, but rather a curated list of links of the best plugins and themes for fish.